The uninformed search is completed, but the frontier has an error WIP
Added the init file(solver.py) that reads from the cmd line.

Implementation questions:
1. Put the Frontier and the Node classes in the datastructures.py, or
pass them to the uninformed_search.py, as it is, the search will have
external dependencies, although these make some sense.
2. The actions functions is generating Vertices(id, weight), instead of
just the id as a string. This clears the code when creating the child nodes,
but it might be faster to just use strings and search the dictionary.
3. I put the solution function returning actions instead of nodes.
Dunno if its the correct.